依頼のあった箇所を修正
ref: https://github.com/tokyo-metropolitan-gov/covid19/pull/4333#issuecomment-631902245
