+++ components/cards/MonitoringConfirmedCasesNumberCard.vue
@@ -1,0 +1,95 @@
+<template>
+  <v-col cols="12" md="6" class="DataCard">
+    <simple-mixed-chart
+      :title="$t('モニタリング指標(1)新規陽性者数')"
+      :title-id="'number-of-confirmed-cases'"
+      :chart-id="'time-bar-chart-patients'"
+      :chart-data="PatientsGraph"
+      :date="PatientsDate"
+      :items="PatientsItems"
+      :labels="PatientsLabels"
+      :data-labels="PatientsDataLabels"
+      :table-labels="PatientsTableLabels"
+      :unit="$t('人')"
+      :url="'https://catalog.data.metro.tokyo.lg.jp/dataset/t000010d0000000068'"
+    >
+      <template v-slot:additionalNotes>
+        <ul :class="$style.GraphDesc">
+          <li>
+            {{ $t('（注）保健所から発生届が提出された日を基準とする') }}
+          </li>
+          <li>
+            {{ $t('（注）医療機関等が行った検査も含む') }}
+          </li>
+          <li>
+            {{
+              $t('（注）チャーター機帰国者、クルーズ船乗客等は含まれていない')
+            }}
+          </li>
+          <li>
+            {{
+              $t(
+                '（注）集団感染発生や曜日による件数のばらつきにより、日々の結果が変動するため、こうしたばらつきを平準化し全体の傾向を見る趣旨から、過去７日間の移動平均値を折れ線グラフで示す（たとえば、5月7日の移動平均値は、5月1日から5月7日までの実績値を平均したもの）'
+              )
+            }}
+          </li>
+        </ul>
+      </template>
+    </simple-mixed-chart>
+  </v-col>
+</template>
+
+<script>
+import dayjs from 'dayjs'
+import duration from 'dayjs/plugin/duration'
+import SimpleMixedChart from '@/components/SimpleMixedChart.vue'
+import Data from '@/data/daily_positive_detail.json'
+dayjs.extend(duration)
+
+export default {
+  components: {
+    SimpleMixedChart
+  },
+  data() {
+    // 検査実施日別状況
+    const l = Data.data.length
+    const PatientsCount = []
+    const sevendayMoveAverages = []
+    const PatientsLabels = []
+    for (let i = 0; i < l; i++) {
+      PatientsCount.push(Data.data[i].count)
+      sevendayMoveAverages.push(Data.data[i].weekly_average_count)
+      PatientsLabels.push(Data.data[i].diagnosed_date)
+    }
+
+    const PatientsGraph = [PatientsCount, sevendayMoveAverages]
+    const PatientsItems = [this.$t('陽性者数'), this.$t('７日間移動平均')]
+    const PatientsDataLabels = [this.$t('陽性者数'), this.$t('７日間移動平均')]
+    const PatientsTableLabels = [this.$t('陽性者数'), this.$t('７日間移動平均')]
+
+    const PatientsDate = Data.date
+    const data = {
+      PatientsDate,
+      PatientsGraph,
+      PatientsItems,
+      PatientsLabels,
+      PatientsDataLabels,
+      PatientsTableLabels
+    }
+    return data
+  }
+}
+</script>
+
+<style module lang="scss">
+.Graph {
+  &Desc {
+    margin: 0;
+    margin-top: 1rem;
+    padding-left: 0 !important;
+    color: $gray-3;
+    list-style: none;
+    @include font-size(12);
+  }
+}
+</style>
